TAG224 2020/06/01   Core     ・YM2151,YM2609,YMF262:Yコマンドを使用するとキャストに失敗するバグを修正。     (さしゅさんからご指摘いただきました。ありがとうございます！)

+++ mml2vgm/Core/chips/YM2151.cs
@@ -885,8 +885,8 @@ namespace Core
             }
             else
             {
-                byte adr = (byte)mml.args[0];
-                byte dat = (byte)mml.args[1];
+                byte adr = (byte)(int)mml.args[0];
+                byte dat = (byte)(int)mml.args[1];
                 parent.OutData(mml, port[0], adr, dat);
             }
         }
